+++ src/actionCreator.js
@@ -1,6 +1,6 @@
 /* @flow strict-local */
 import type { Action, GlobalState, Dispatch } from './types';
-import { clearTypingNotification } from './typing/clearTypingNotification';
+import { typingStatusExpiryLoop } from './typing/typingActions';
 import { EVENT_TYPING_START } from './actionConstants';
 
 export default (dispatch: Dispatch, actions: $ReadOnlyArray<Action>, state: GlobalState) => {
@@ -8,7 +8,7 @@ export default (dispatch: Dispatch, actions: $ReadOnlyArray<Action>, state: Glob
     switch (action.type) {
       case EVENT_TYPING_START:
         if (Object.keys(state.typing).length === 0) {
-          dispatch(clearTypingNotification());
+          dispatch(typingStatusExpiryLoop());
         }
         break;
       default:
